feat(info): finished
